Added a test scene and the basic o2 system
The spaceship test scene is to test the different systems until we get a
concrete spaceship. Until than we can build our own spaceship and use it
to test the different systems.

I added a basic o2 system through doors. Each room is connected by
doors. If the door is open it tries to get the same amount of oxygen in
both rooms through adding and deducting oxygen.