Added a way to interact with ingame objects
Every object which wants to be interactable should implement
IActivatable and should have a trigger. The player will look when he
enters a trigger if that collider has a IActivatable Component. If it
has one it will be used for all interactions.

I'm not 100% happy but it's good enough for now.